move FacetIndex into facet_index.rs
